Many Improvements and Bugfixes
/NICK NewNickname command if there is a nick NewNickname nick
on a channel the command PARTS from that channel

Fixed up the joined and joining dicts. They are storing #channel
records now.

/NAMES command and a nickname list which
is updated on any JOIN and PART.

On JOIN the ircd ignores the first presence with the own nickname but
without any x content.

WEBIRC command which is useful for webchats. Of course this
does not make any sense yet until there is a maximum clients from one
IP limit for the ircd connection.

/LIST command which displays all the cannels configured in
the ejabberd.conf

Committer: Hans Freitag <zem@fnordpol.de>